+++ src/i18n/index.js
@@ -1,12 +1,8 @@
 import i18n from 'i18next'
 import { initReactI18next } from 'react-i18next'
-import Backend from 'i18next-xhr-backend'
-import LanguageDetector from 'i18next-browser-languagedetector'
 import en from './en'
 
 i18n
-  .use(Backend)
-  .use(LanguageDetector)
   .use(initReactI18next) // passes i18n down to react-i18next
   .init({
     lng: 'en',