+++ src/utils/boundingBoxSelection/setBoundingBoxNodes.js
@@ -3,6 +3,7 @@ import { ROUTE_NETWORK_GRAPHS } from '../../constants/routes'
 import { OPERATION_TYPE_UPDATE } from '../../constants/store'
 import { SIDEBAR_VIEW_GRAPHS } from '../../constants/views'
 import store from '../../store'
+import setPageView from '../analytics/setPageView'
 
 /**
  * Set nodes inside bounding box
@@ -56,6 +57,7 @@ const setBoundingBoxNodes = ({
   updateStoreValue(['lastGraphIndex'], OPERATION_TYPE_UPDATE, newGraphIndex)
   updateStoreValue(['sidebarView'], OPERATION_TYPE_UPDATE, SIDEBAR_VIEW_GRAPHS)
   window.history.pushState('', '', ROUTE_NETWORK_GRAPHS)
+  setPageView({ url: ROUTE_NETWORK_GRAPHS, updateStoreValue })
 }
 
 export default setBoundingBoxNodes