+++ src/utils/shortestPath/setShortestPath.js
@@ -4,6 +4,7 @@ import store from '../../store'
 import { OPERATION_TYPE_UPDATE } from '../../constants/store'
 import { ROUTE_NETWORK_GRAPHS } from '../../constants/routes'
 import { SIDEBAR_VIEW_GRAPHS } from '../../constants/views'
+import setPageView from '../analytics/setPageView'
 
 /**
  * Set shortest path
@@ -76,6 +77,7 @@ const setShortestPath = async ({
   updateStoreValue(['lastGraphIndex'], OPERATION_TYPE_UPDATE, newGraphIndex)
   updateStoreValue(['sidebarView'], OPERATION_TYPE_UPDATE, SIDEBAR_VIEW_GRAPHS)
   window.history.pushState('', '', ROUTE_NETWORK_GRAPHS)
+  setPageView({ url: ROUTE_NETWORK_GRAPHS, updateStoreValue })
 }
 
 export default setShortestPath